version working (just remains to create the correct results.dat after the survey).

+++ Template/LO/SubProcesses/survey.sh
@@ -6,23 +6,38 @@ k=run1_app.log
 script=ajob1                         
 offset=$1
 shift
+subdir=$offset
 for i in $@ ; do
      j=G$i
-     echo $offset  > moffset.dat
-     if [[ $offset -gt 0 ]]; then
-	 j=G${i}_${offset}
+     if [[ $offset == *.* ]];then
+	 subdir=${offset%.*}
+	 offset=${offset##*.}
+	 j=G${i}_${subdir}	     
+     elif [[ $offset -gt 0 ]]; then
+	 j=G${i}_${subdir}
      fi
      if [[ ! -e $j ]]; then
          mkdir $j
      fi
      cd $j
-     rm -f ftn25 ftn26 ftn99
-     rm -f $k
-     cat ../input_app.txt >& input_app.txt
+     if [[ $offset -eq 0 ]]; then
+	 rm -f ftn25 ftn26 ftn99
+	 rm -f $k
+     else
+	 echo   "$offset"  > moffset.dat
+     fi
+     if [[ $offset -eq $subdir ]]; then
+	 rm -f ftn25 ftn26 ftn99
+	 rm -f $k
+     fi
+     if [[ ! -e input_app.txt ]]; then 
+	 cat ../input_app.txt >& input_app.txt
+     fi
      echo $i >> input_app.txt
+
      for((try=1;try<=10;try+=1)); 
      do
-     ../madevent > $k <input_app.txt
+     ../madevent >> $k <input_app.txt
      if [ -s $k ]
      then
          break
@@ -30,8 +45,8 @@ for i in $@ ; do
      sleep 1
      fi
      done
-     rm -f ftn25 ftn99
-     rm -f ftn26
+#     rm -f ftn25 ftn99
+#     rm -f ftn26
      echo "" >> $k; echo "ls status:" >> $k; ls >> $k
      cp $k log.txt
      cd ../
